Handle metadata from pulp-admin CLI
This commit allows users to specify some glance metadata via pulp-admin and
have it passed through to pulp. Some options are in here but I need to sync up
with the glance team to decide on a full list of options.

Pulp-admin will constrain what data can be sent but calling via REST allows
arbitrary metadata. Also, the 'protected' flag is set when images are uploaded
via pulp-admin.