Publish images to glance
This creates a new glance distributor as well as changes to pulp-admin.
Configuration of the distributor is done at repo creation time via command-line
arguments for now.

If an image already exists in glance, it is skipped. Images are correlated
between glance and pulp via image checksum and repo names. Additional metadata
is added to the image file in glance to make it easy for pulp to find.