Support for deleting images
If an image is in glance and associated with a repo yet was not touched by pulp
during the publish, it will be removed.

There is a failsafe to bail out totally if any images are in pulp that should
have been published and weren't.

The finalize() method could use some refactoring at some point in the future.
It is a little overly verbose right now.