remove pep257 requirement from tests
No other plugin uses this, and it's currently breaking due to a change
in pep257 that would require yet another exception to be added. As far
as I can tell, we've added exceptions for every error that pep257 might
tell us about, so this change is effectively a no-op.

I do think docstrings are good, and that we should have them, so I'm
referring to a story that I just made about getting docstrings added to
pulp. Discussions of the pro/cons of requiring docstrings should
probably happen there.

Furthermore, this commit can be used as a reference for seeing how we
integrated pep257 once upon a time.

re #1718
https://pulp.plan.io/issues/1718